add IndexTypeCacheProvider
