+++ src/YesSql.Abstractions/Indexes/IIndexTypeCacheProvider.cs
@@ -1,0 +1,12 @@
+using System;
+using System.Reflection;
+using YesSql.Serialization;
+
+namespace YesSql.Indexes
+{
+    public interface IIndexTypeCacheProvider
+    {
+        public PropertyInfo[] GetTypeProperties(Type type);
+        public PropertyInfoAccessor GetPropertyAccessor(PropertyInfo property);
+    }
+}
