+++ src/YesSql.Abstractions/Serialization/PropertyInfoAccessor.cs
